Implement getting bool configuration values
GetBool provides a cleaner interface for using binary options than
making string comparisons. Because there isn't really a "safe" failure
value for bool, the user is required to pass in the default value. I.e.:
"If the value isn't set or can't be retrieved, this is the value I
want". This should be a reasonable balance between convenience and
safety.