Read config.secret.json for any config load
This allows any and all secrets to be stored in `config.secret.json` and
because it is loaded last, it overrides existing values loaded from
another file (if needed). This file is loaded any time a config is
loaded so that the filename never needs to be explicitly specified when
loading a config.

Also remove unneeded references to `config.secret.json` in the password
init.