Fix MongoDB data injection vulnerability
Update the go.mongodb.org/mongo-driver module to fix the vulnerability.
This hasn't been much of a risk as Mongo is not actively used, however,
the code still exists in the repo, thus it's better to fix now.

Further reading: https://github.com/advisories/GHSA-f6mq-5m25-4r72 or
CVE-2021-20329